+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Update fee/fine actions column UX for accessibility. Refs UIU-3027.
 * Change import of `exportToCsv` from `stripes-util` to `stripes-components`. Refs UIU-3202.
 * Provide `itemToString` to create unique `key` attributes. Refs UIU-3312.
+* Return appropriate error message when item for manual fee/fine can't be found. Refs UIU-2701.
 
 ## [11.0.11](https://github.com/folio-org/ui-users/tree/v11.0.11) (2025-01-15)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v11.0.10...v11.0.11)